requirements.txt
Pillow>=10.2.0
piexif==1.1.3 